+++ src/ErrorEvents.ts
@@ -2,5 +2,5 @@ import type { Document, EventsDescription } from "mongodb";
 
 export interface ErrorEvents<TMessage extends Document>
   extends EventsDescription {
-  error: (err: Error, message?: TMessage) => unknown;
+  error: (err: Error, message?: TMessage) => void;
 }